Extract LambdaCompiler
